refactor: change paths to data files in elements
- for Button
- for Form
- for Input
- for Label
