+++ src/features/input/elements/Button/Button.js
@@ -1,4 +1,4 @@
-import data from './button.json'
+import data from './Button.data.json'
 import createCustomizedElement from '../../utils/createCustomizedElement'
 
 const Button = createCustomizedElement(data.name, data.attrs, data.children)
