+++ src/features/input/elements/Input/Input.js
@@ -1,4 +1,4 @@
-import data from './input.json'
+import data from './Input.data.json'
 import createCustomizedElement from '../../utils/createCustomizedElement'
 
 const Input = createCustomizedElement(data.name, data.attrs, data.children)
