+++ src/features/input/elements/Label/Label.js
@@ -1,4 +1,4 @@
-import data from './label.json'
+import data from './Label.data.json'
 import createCustomizedElement from '../../utils/createCustomizedElement'
 
 const Label = createCustomizedElement(data.name, data.attrs, data.children)
